splat: make aux_precision:half the default (model-level verdict overrides the pixel-level gate)
868e17f kept `aux_precision: half` opt-in-only because it fails this repo's
original tight pixel-level bar (MV median error 0.012-0.016px vs. a
<=0.01px bar, depth relative median error ~5.4e-3 vs. a <=1e-3 bar,
measured on demo/data/clip_0000.npz). The prior commit on this branch
fixed the real bug that made "half" unsafe at any bar -- out_fg's blend
alpha breaking under RG16F -- so "half" was already correctness-safe, just
imprecise relative to that pixel-level bar.

Coordinator's model-level verdict changes the calculus: the actual
mobiledlss DLSS reconstructor, trained/evaluated on lux-rendered clips
(not a pixel-exact .npy diff), finds "half" indistinguishable from
"float": 35.494 vs. 35.491 dB PSNR, fg bit-exact, MV 0.008px, depth 0.4%
-- all invisible to the reconstructor, which only ever consumes
out_aux/out_fg through further lossy stages (bilinear resampling, a
learned network) with far more slack than a pixel-exact comparison
allows. Since the pixel-level bar was never the actual product
requirement -- reconstruction quality is -- flip the default to "half".

Changes:
- luxc/expansion/splat_expander.py: default "aux_precision" -> "half",
  rewrote the option's docstring to lead with the model-level verdict.
- luxc/codegen/reflection.py: matching fallback default -> "half"
  (defensive; splat_expander.py always populates the key today).
- Regenerated examples/gaussian_splat_dlss.comp.json (the only DLSS
  example besides gaussian_splat_dlss_half.lux, per
  `grep -l motion_vectors examples/*.lux`) via
  `python -m luxc examples/gaussian_splat_dlss.lux` -- SPIR-V byte-
  identical, only the reflected "aux_precision" field flips (host format
  hint only, doesn't affect codegen).
- examples/gaussian_splat_dlss_half.lux: now functionally identical to
  gaussian_splat_dlss.lux's default; kept (not deleted) as an explicit,
  self-documenting example of setting `aux_precision: half` by name, per
  the coordinator's "keep as an explicit example or delete" -- updated
  its header comment to explain the redundancy and cite the model-level
  numbers.
- SPECIFICATION.md, docs/language-reference.md, docs/rendering-engines.md:
  updated to describe "half" as the default, "float" as the still-
  selectable pixel-precise option, and the model-level justification.
- tests/test_dlss_outputs.py: PIPELINE_BASE (examples/gaussian_splat_dlss)
  now defaults to "half", so two of its pixel-exact assertions needed
  loosening (both cite the model-level result inline):
    * TestExpectedDepth::test_single_opaque_splat_depth_equals_camera_z:
      1e-3 absolute -> 1e-2 *relative* (single-splat half-float depth
      rounding measures ~0.07% relative here, well inside the new bound;
      the old absolute bound was ~1.8e-4 relative at this test's depth).
    * TestJitter::test_jitter_does_not_affect_motion_vectors: 1e-3px ->
      0.02px (two independent half-float-rounded renders differ by
      ~0.003px here, well inside the new bound).
  All other tests in the file (including the mv=0-by-construction case,
  where half-float rounds 0.0 to 0.0 exactly) were already unaffected.

Verification (Mac, both binaries rebuilt with the out_fg fix from the
prior commit -- no further rebuild needed here, since aux_precision is a
runtime host format hint read from the compiled JSON, not compiled into
the binaries):
- tests/test_splat_dlss_outputs.py: 8 passed.
- tests/test_dlss_outputs.py: 20 passed (after the two tolerance
  loosenings above).
- Full lux-4dgs suite (`pytest tests/ -q`): 1559 passed, 26 skipped,
  0 failed (946s).
- mobiledlss tests/test_lux_reconstruct.py, both binaries
  (LUX_PLAYGROUND=.../build/lux-playground and
  LUX_PLAYGROUND=.../build-metal/lux-playground-metal): 3 passed each
  (unrelated to aux_precision -- reconstruct-stage pipeline exactness --
  run as a rebuild-regression check per the coordinator's request).
- Re-measured true float-vs-half (via a scratch aux_precision:float
  pipeline, since examples/gaussian_splat_dlss is now half by default)
  on demo/data/clip_0000.npz's target camera, juggle_full.glb, both
  binaries:
    frame 0 (trivial, mv=0 by construction): fg bit-exact on both
      backends, MV median diff ~6e-7px, depth relative median diff
      ~0.534%
    frame 1 (real actor+camera motion): fg bit-exact on both backends,
      MV median diff ~0.021px, depth relative median diff ~0.537%
  (fg's bit-exactness reconfirms the prior commit's out_fg fix; MV/depth
  numbers are consistent with the model-level report's ballpark.)
- Mac Metal CLI timing, examples/gaussian_splat_dlss (now half by
  default) at 960x540, dynamic DLSS scene
  (demo/ios_assets/juggle_p0.8_stride4.glb, frame 10, sort view_depth,
  --splat-backend luxc default): `render(): preprocess=0.026ms
  sort=0.033ms render=5.46-5.50ms` across 3 runs -- consistent with
  868e17f's documented "half and float indistinguishable, ~5.6-5.9ms"
  finding (M4 Max isn't attachment-bandwidth-bound at this scale).

playground_ios/ and playground_android_perf/ have other uncommitted work
in this worktree, left untouched -- out of scope for this change.

+++ luxc/codegen/reflection.py
@@ -469,12 +469,14 @@ def generate_reflection(
             "expected_depth": splat_config.get("expected_depth", False),
             "foreground_coverage": splat_config.get("foreground_coverage", False),
             # Host format hint for out_aux ONLY (bench/lux_perf_ablation.md
-            # task 2 follow-up) -- "float" (out_aux RGBA32F, precision-safe,
-            # default) or "half" (out_aux RGBA16F, smaller but measured to
-            # fail the parity gate -- see splat_expander.py's docstring).
-            # out_fg (when present) always stays RGBA16F, independent of
-            # this setting -- see splat_expander.py's docstring for why.
-            "aux_precision": splat_config.get("aux_precision", "float"),
+            # task 2 follow-up) -- "half" (out_aux RGBA16F, DEFAULT: fails
+            # the original tight pixel-level bar but is model-level
+            # indistinguishable from "float", see splat_expander.py's
+            # docstring) or "float" (out_aux RGBA32F, the pixel-precise
+            # option, still selectable). out_fg (when present) always
+            # stays RGBA16F, independent of this setting -- see
+            # splat_expander.py's docstring for why.
+            "aux_precision": splat_config.get("aux_precision", "half"),
         }
         if stage.stage_type == "compute":
             gs_meta["workgroup_size"] = [256, 1, 1]
